Introduce new HackAndSlash
